+++ lib/exercises_index.js
@@ -1,6 +1,7 @@
 var jQuery = require('jquery');
 window.$ = jQuery;
-var exercise = require('./exercise.js')
+var exercise = require('./exercise.js');
+var exerciseTable = require('./exercises_table.js');
 
 function populateExercises(){
   var currentExercises = localStorage.getItem('exercises');
@@ -8,15 +9,10 @@ function populateExercises(){
     var currentExercisesJSON = JSON.parse(currentExercises);
     for (var i = 0; i < currentExercisesJSON.length; i++) {
       var exerciseItem = new exercise(currentExercisesJSON[i]['name'], currentExercisesJSON[i]['calories']);
-      appendToExercises(exerciseItem);
+      exerciseTable.appendTo(exerciseItem);
     }
   }
 }
-
- function appendToExercises(exerciseItem){
-   var deleteButton = '<button class="delete-button"><b>-</b></button>'
-   $('#exercises-table tr:first').after('<tr><td class="exercise-cell exercise-name-cell">' + exerciseItem.name + '</td><td class="exercise-cell exercise-calorie-cell">' + exerciseItem.calories + '</td><td class="delete-cell">' + deleteButton + '</td></tr>')
- }
 
  function clearExerciseForm(){
    $('#add-exercise-form input[name="name"]').val('');
@@ -47,35 +43,6 @@ function populateExercises(){
    $('#calorie-error').addClass('hidden');
  }
 
- function filterExercises() {
-   var input = document.getElementById('exercise-filter');
-   var filter = input.value.toUpperCase();
-   var table = document.getElementById('exercises-table');
-   var tr = table.getElementsByTagName("tr");
-
-   for (i = 0; i < tr.length; i++) {
-     td = tr[i].getElementsByTagName("td")[0];
-     if (td) {
-       if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
-         tr[i].style.display = "";
-       } else {
-         tr[i].style.display = "none";
-       }
-     }
-   }
- }
-
- function updateCell(cell, originalContent, newContent){
-   if(cell.hasClass('exercise-name-cell')){
-     var exerciseItem = exercise.find(originalContent);
-     exerciseItem.update('name', newContent);
-   } else if(cell.hasClass('exercise-calorie-cell')){
-     var exerciseName = cell.siblings('.exercise-name-cell').html();
-     var exerciseItem = exercise.find(exerciseName);
-     exerciseItem.update('calories', newContent);
-   }
- }
-
 $(document).ready(function(){
   populateExercises();
 
@@ -91,7 +58,7 @@ $(document).ready(function(){
     } else {
       exerciseItem = new exercise(name, calories)
       exerciseItem.store();
-      appendToExercises(exerciseItem);
+      exerciseTable.appendTo(exerciseItem);
       clearExerciseForm();
       removeErrors();
     }
@@ -104,7 +71,7 @@ $(document).ready(function(){
   });
 
   $('#exercise-filter').keyup(function(){
-    filterExercises();
+    exerciseTable.filter();
   })
 
   $('#exercises-table').on('click', '.exercise-cell', function(e){
@@ -123,7 +90,7 @@ $(document).ready(function(){
       if(e.which == 13 || e.type == 'blur') {
         var newContent = $(this).val();
         $(this).parent().text(newContent);
-        updateCell(cell, originalContent, newContent);
+        exerciseTable.updateCell(cell, originalContent, newContent);
       }
     });
   });
