+++ lib/exercises_table.js
@@ -1,0 +1,38 @@
+var exercise = require('./exercise.js');
+var ExerciseTable = function(){}
+
+ExerciseTable.appendTo = function(exerciseItem){
+  var deleteButton = '<button class="delete-button"><b>-</b></button>'
+  $('#exercises-table tr:first').after('<tr><td class="exercise-cell exercise-name-cell">' + exerciseItem.name + '</td><td class="exercise-cell exercise-calorie-cell">' + exerciseItem.calories + '</td><td class="delete-cell">' + deleteButton + '</td></tr>');
+}
+
+ExerciseTable.filter = function(){
+  var input = document.getElementById('exercise-filter');
+  var filter = input.value.toUpperCase();
+  var table = document.getElementById('exercises-table');
+  var tr = table.getElementsByTagName("tr");
+
+  for (i = 0; i < tr.length; i++) {
+    td = tr[i].getElementsByTagName("td")[0];
+    if (td) {
+      if (td.innerHTML.toUpperCase().indexOf(filter) > -1) {
+        tr[i].style.display = "";
+      } else {
+        tr[i].style.display = "none";
+      }
+    }
+  }
+}
+
+ExerciseTable.updateCell = function(cell, originalContent, newContent){
+  if(cell.hasClass('exercise-name-cell')){
+    var exerciseItem = exercise.find(originalContent);
+    exerciseItem.update('name', newContent);
+  } else if(cell.hasClass('exercise-calorie-cell')){
+    var exerciseName = cell.siblings('.exercise-name-cell').html();
+    var exerciseItem = exercise.find(exerciseName);
+    exerciseItem.update('calories', newContent);
+  }
+}
+
+module.exports = ExerciseTable;
